requirements.txt
anthropic>=0.25.0
pydantic>=2.0.0
pyyaml>=6.0
rich>=13.0.0
typer>=0.9.0
python-dateutil>=2.8.2
jinja2>=3.1.2
pytest>=7.4.0
